removed the interactive features
