+++ packages/core/src/plugins/index.ts
@@ -1,3 +1,2 @@
 export * from './loader';
 export * from './bundle';
-export * from './manager';
