+++ packages/fastify/src/index.ts
@@ -4,7 +4,6 @@ import flugin from 'fastify-plugin';
 import { loadPlugins } from '@zotera/core';
 
 import { configDecorator } from './decorators/config';
-import { pluginsDecorator } from './decorators/plugins';
 import { storageDecorator } from './decorators/storage';
 import { routes } from './routes';
 import { ping } from './routes/ping';
@@ -13,17 +12,20 @@ import type { PluginOptions } from './types';
 const plugin: FastifyPluginCallback<PluginOptions> = flugin(
   async (app: FastifyInstance, options: PluginOptions, next: (error?: FastifyError) => void) => {
     app.register(configDecorator, options);
-    app.register(pluginsDecorator);
 
     // Setup plugin loading
     loadPlugins(options);
 
+    // This needs to be after the plugins are loaded
+    // so that the plugins that are a storage plugin
+    // will be available to the storage decorator
+    app.register(storageDecorator);
+
     app.register(ping, {
       prefix: '/-/ping'
     });
-    app.register(storageDecorator);
-    app.register(routes);
 
+    app.register(routes);
     next();
   },
   {
@@ -32,3 +34,4 @@ const plugin: FastifyPluginCallback<PluginOptions> = flugin(
 );
 
 export default plugin;
+
